+++ ChangeLog.md
@@ -50,10 +50,12 @@
 - **bin**
   
   - +) `StructToMap` 和 `StructToMapLStyle` 新增 struct 转 map 的方法
+  - +) `CliCmd.AppendData` 新增数据合并接口，使 CliCmd 支持外部数据导入。
   - 优化) `FormatQue` 使用反射调整其支持任一类型的 array/slice数据
   
 - **util**
   - +) `InQueAny` 新增方法来支持任一类型的 array/slice数据键值检测
+  - +) `SendTimeDiff` 新增方法用于计算更加统计的时间消耗，用于替代旧方法：`Sec、SecCall、SecCallStr`。
   
 - **str**
   - +) `IsLatinAlpha` 判断是否为拉丁文字母
@@ -64,6 +66,10 @@
   - +) `GetMemStatsRange` 新增方法用于计算程序内存消耗
   - +) `MemUsage` 新增结构体用于快速计算系统的内存消耗情况
 
+- **cmd/uymas**
+  
+  - optimize) `scan, sc` 增加内存使用监控、时间消耗的监控。
+  
   
 
 ### 1.0.1/2021-06-11